+++ crates/gitbutler-command-context/src/lib.rs
@@ -98,18 +98,3 @@ impl CommandContext {
         )?)
     }
 }
-
-// TODO(ST): put this into `gix`, the logic seems good, add unit-test for number generation.
-pub trait GixRepositoryExt: Sized {
-    /// Configure the repository for diff operations between trees.
-    /// This means it needs an object cache relative to the amount of files in the repository.
-    fn for_tree_diffing(self) -> Result<Self>;
-}
-
-impl GixRepositoryExt for gix::Repository {
-    fn for_tree_diffing(mut self) -> anyhow::Result<Self> {
-        let bytes = self.compute_object_cache_size_for_tree_diffs(&***self.index_or_empty()?);
-        self.object_cache_size_if_unset(bytes);
-        Ok(self)
-    }
-}
